Refactoring towards

+++ src/main/java/uz/jamshid/apelsin/payload/HighDemandProductDto.java
@@ -4,10 +4,12 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigInteger;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class HighDemandProductDto {
     private Integer productId;
-    private Short quantity;
+    private BigInteger quantity;
 }
