+++ src/main/java/uz/jamshid/apelsin/payload/NumberOfProductsInYearDto.java
@@ -4,10 +4,12 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigInteger;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class NumberOfProductsInYearDto {
-    private Integer orderId;
+    private BigInteger orderId;
     private String country;
 }
